+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-## 1.2.0 / unreleased
+## 1.2.0 / 2019-12-08
 
 **NOTE** For security reason, the Web UI is disabled by default. In order to enable it, pass the `--web.ui-enabled` flag
 when program starts.